+++ src/fe_romlist.hpp
@@ -210,7 +210,7 @@ public:
 		std::vector< std::pair<std::string, bool> > &tags_list ) const;
 	bool set_tag( FeRomInfo &rom, FeDisplayInfo &display, const std::string &tag, bool flag );
 
-	int filter_size( int filter_idx ) const { return ( filter_idx < m_filtered_list.size() ) ? (int)m_filtered_list[filter_idx].filter_list.size() : 0; };
+	int filter_size( int filter_idx ) const { return ( filter_idx < (int)m_filtered_list.size() ) ? (int)m_filtered_list[filter_idx].filter_list.size() : 0; };
 	const FeRomInfo &lookup( int filter_idx, int idx) const { return *(m_filtered_list[filter_idx].filter_list[idx]); };
 	FeRomInfo &lookup( int filter_idx, int idx) { return *(m_filtered_list[filter_idx].filter_list[idx]); };
